Get and filter based on only part of query
The expression featured === "true" ? true : false
checks if featured is exactly the string "true".
If it is, queryObject.featured is set to true, otherwise
it is set to false. his means that queryObject.featured
will always be a boolean (true or false), regardless
of the original value of featured.
We did not do queryObject.featured = featured directly
because if featured is a string (e.g., "true" or "false"),
then queryObject.featured will also be that string, not a boolean.